hashing out first instance of new relationship stuff
Still missing some code to make has_one truely unique and such but the general
idea is working, it seems.

add `documents` property to RethinkCollection which will cache the results or
fetch them, just as a helper.

test.py was/is some fake code that happens to actually work(ish) with
RethinkRelation